已修复。
问题原因：

bf16-mixed 下 spike_rate 张量是 BFloat16，但 spike_rate_valid 来自 SNN 统计是 Float32，索引赋值要求 dtype 完全一致。
修复内容：

在赋值前强制转换：
spike_rate[hist_feat_key_valid] = spike_rate_valid.to(dtype=spike_rate.dtype)

+++ src/model/model_forecast_snn_fast.py
@@ -83,7 +83,7 @@ class SNNModelForecastFast(ModelForecast):
         )
         spike_rate = torch.zeros(B * N, device=actor_feat_valid.device, dtype=actor_feat_valid.dtype)
         actor_feat[hist_feat_key_valid] = actor_feat_valid
-        spike_rate[hist_feat_key_valid] = spike_rate_valid
+        spike_rate[hist_feat_key_valid] = spike_rate_valid.to(dtype=spike_rate.dtype)
         actor_feat = actor_feat.view(B, N, -1)
         spike_rate = spike_rate.view(B, N)
         return actor_feat, spike_rate, self.temporal_encoder.latest_aux_losses
